64-bit compatibility

+++ src/demofile.h
@@ -44,7 +44,7 @@ typedef int32_t         int32;
 typedef uint32_t        uint32;
 typedef int64_t         int64;
 typedef uint64_t        uint64;
-typedef unsigned long   CRC32_t;
+typedef uint32_t        CRC32_t;
 
 // Demo messages
 enum
